refactor: modularize inventory processor
